requirements.txt
# Fetch.ai uAgents Framework
uagents>=0.12.0
uagents-core>=0.1.0

# SingularityNET MeTTa
hyperon>=0.1.0

# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging
colorlog>=6.7.0

# Utilities
python-dateutil>=2.8.2
